dev/tools/tse-reporter/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
click>=8.1.0
rich>=13.0.0
